requirements.txt
tensorflow>=2.16
keras>=3.0
numpy>=1.26
pandas>=2.2
scikit-learn>=1.4
tqdm>=4.66
requests>=2.32
pillow>=10.3
opencv-python>=4.10
matplotlib>=3.9